fixes to osx version, updates to bounds dialog box, add 2d plots for histograms and devices, bug fixes -

+++ Build/smokeview/intel_osx_64/make_smokeview.sh
@@ -1,9 +1,26 @@
 #!/bin/bash
-OPTS="$*"
-OPTS="-q $*"
+if [ "$NOQUARTZ" != "" ]; then
+  OPTS="-Q $*"
+  if [ "$LOWRES" != "" ]; then
+    echo building smokeview no Quartz, low res
+  else
+    echo building smokeview no Quartz, high res
+  fi
+else
+  echo building smokeview using Quartz
+  OPTS="-q $*"
+fi
 source ../../scripts/setopts.sh $OPTS
 
-LIBDIR=../../LIBS/intel_osx_64
+if [ "$QUARTZSMV" == "use_quartz" ]; then
+  LIBDIR=../../LIBS/intel_osx_64
+else
+  if [ "$LOWRES" == "" ]; then
+    LIBDIR=../../LIBS/intel_osx_noq_64
+  else
+    LIBDIR=../../LIBS/intel_osx_noql_64
+  fi
+fi
 
 CURDIR=`pwd`
 cd $LIBDIR
@@ -22,12 +39,16 @@ do
     build_libs=1
   fi
 done
+if [ "$BUILD_LIBS" == "1" ]; then
+  build_libs=1
+fi
 if [ "$build_libs" == "1" ]; then
-  cd $LIBDIR
   echo building libs
   eval ./make_LIBS.sh $OPTS
 fi
 cd $CURDIR
 
-rm -f *.o *.mod smokeview_osx_64*
+if [ "$BUILD_ALL" == "1" ]; then
+  rm -f *.o *.mod smokeview_osx_64*
+fi
 eval make -j 4 QUARTZ="$QUARTZSMV" GLUT="$GLUT" ${SMV_MAKE_OPTS} -f ../Makefile intel_osx_64
